odbcBackend: Rename plugin->pass kdb plugin-check
- rename from backendOdbc to backend_odbc
- add entry in README.md file for plugins
