+++ src/libs/elektra/kdb.c
@@ -919,14 +919,14 @@ static bool addHardcodedMountpoints (KDB * handle, Key * errorKey)
 
 
 	/* FIXME: Added for testing purpose only (during dev), must be removed before merging into master! */
-	Plugin * backendOdbc = elektraPluginOpen ("backendOdbc", handle->modules, ksNew (0, KS_END), errorKey);
-	if (backendOdbc == NULL)
+	Plugin * backend_odbc = elektraPluginOpen ("backend_odbc", handle->modules, ksNew (0, KS_END), errorKey);
+	if (backend_odbc == NULL)
 	{
 		ELEKTRA_SET_INSTALLATION_ERROR (errorKey, "Could not open system:/elektra/odbc backend. See warnings for details.");
 		return false;
 	}
-	backendOdbc->global = handle->global;
-	addMountpoint (handle->backends, keyNew (KDB_SYSTEM_ELEKTRA "/odbc", KEY_END), backendOdbc, ksNew (0, KS_END), ksNew (0, KS_END));
+	backend_odbc->global = handle->global;
+	addMountpoint (handle->backends, keyNew (KDB_SYSTEM_ELEKTRA "/odbc", KEY_END), backend_odbc, ksNew (0, KS_END), ksNew (0, KS_END));
 
 	/* FIXME: End of testing code to be removed */
